feat: 플레이어 인스턴스 바인딩 담당하는 PlayerInstanceManager 추가

+++ Assets/_Game/UI/Scripts/HealthBarHudUI.cs
@@ -35,7 +35,7 @@ namespace DungeonShooter
         }
         private void UpdateVisuals()
         {
-            _targetFillAmount = (float)_currentHealth / _maxHealth;
+            _targetFillAmount = (float)_currentHealth / (float)_maxHealth;
             
             if (_healthText != null)
                 _healthText.text = $"{_currentHealth} / {_maxHealth}";
